fix(client):  remove sync suggestion from signin view
Fix #3903
