+++ app/tests/spec/views/sign_up.js
@@ -242,6 +242,32 @@ define(function (require, exports, module) {
       });
     });
 
+    describe('sync suggestion - signup-mixin', function () {
+
+      it('displays sync suggestion message if no migration', function () {
+        createView();
+        relier.set('service', null);
+
+        sinon.spy(view.metrics, 'logEvent');
+
+        return view.render()
+          .then(function () {
+            assert.lengthOf(view.$('#suggest-sync'), 1);
+            assert.equal(view.$('#suggest-sync').html(), 'Looking for Firefox Sync? <a href="https://mozilla.org/firefox/sync">Get started here.</a>');
+            assert.isTrue(TestHelpers.isEventLogged(metrics, 'signup.sync-suggest.visible'), 'enrolled');
+          });
+      });
+
+      it('does not display sync suggestion message if there is a relier service', function () {
+        relier.set('service', 'sync');
+
+        return view.render()
+          .then(function () {
+            assert.lengthOf(view.$('#suggest-sync'), 0);
+          });
+      });
+    });
+
     describe('migration', function () {
       it('does not display migration message if no migration', function () {
         return view.render()
